moved questions.js to faq page
