+++ package.json
@@ -46,13 +46,13 @@
     "@storybook/addon-a11y": "3.4.11",
     "@storybook/addon-actions": "3.4.11",
     "@storybook/addon-backgrounds": "3.4.11",
-    "@storybook/addon-info": "3.4.11",
+    "@storybook/addon-info": "^5.1.9",
     "@storybook/addon-knobs": "3.4.11",
     "@storybook/addon-options": "3.4.11",
     "@storybook/addon-storyshots": "3.4.11",
     "@storybook/addon-viewport": "3.4.11",
     "@storybook/addons": "3.4.11",
-    "@storybook/react": "3.4.11",
+    "@storybook/react": "^5.1.9",
     "@zeit/next-bundle-analyzer": "^0.1.2",
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-source-maps": "0.0.4-canary.1",
@@ -97,7 +97,7 @@
     "react-scroll": "^1.7.12",
     "react-scroll-up-button": "^1.6.4",
     "react-select": "^3.0.4",
-    "react-svg-loader": "2.1.0",
+    "react-svg-loader": "^3.0.3",
     "react-youtube": "^7.8.0",
     "reselect": "^4.0.0",
     "rss-parser": "^3.7.1",
