+++ pages/podcasts.js
@@ -39,16 +39,6 @@ class Podcasts extends React.Component {
     error: undefined,
   };
 
-  // Podcasts.defaultProps = {
-  //   children: undefined,
-  //   title: undefined,
-  //   description:
-  // eslint-disable-next-line max-len
-  //     'Interviews with military veterans, spouses, and others from the Operation Code community on their transition to careers to technology',
-  //   url: 'https://operationcode.org',
-  //   ogImage: `${s3}branding/logos/small-blue-logo.png`,
-  // };
-
   render() {
     const { episodes, error } = this.props;
     const pageTitle = 'Podcasts';
